implemented separate pages for homepage and quiz modes
