+++ src/pages/CustomMode.tsx
@@ -1,0 +1,9 @@
+const CustomMode = () => {
+  return (
+    <div>
+      <h1>Custom Quiz </h1>
+    </div>
+  );
+};
+
+export default CustomMode;
